Refactored app.db out of trackers
