+++ bolero/trackers/__init__.py
@@ -1,1 +1,4 @@
+from flask_sqlalchemy import SQLAlchemy
 __all__ = ['twitter', 'wunderlist', 'withings_tracker', 'myfitnesspal_tracker']
+
+db = SQLAlchemy()
